requirements.txt
torch==1.10.2+cu102 -f https://download.pytorch.org/whl/cu102/torch_stable.html
pyyaml
tqdm
h5py
numpy